chore: move `useAnalytics` to `client/hooks`
