+++ apps/meteor/client/hooks/useAnalytics.ts
